player healthbar

+++ Insurgent/Assets/Scripts/Enemy.cs
@@ -1,16 +1,25 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+//using UnityEngine.UI;
+
 
 public class Enemy : MonoBehaviour
 {
-
-    public int health = 100;
+    public float health = 100;
     public GameObject DeathEffect;
+
+    //Image healthBar;
+
+    //void Start()
+    //{
+    //    healthBar = GetComponent<Image>();
+    //}
 
     public void TakeDamage(int damage)
     {
         health -= damage;
+        //healthBar.fillAmount = health / 100;
 
         if (health <= 0)
         {
